% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{available_tables}
\alias{available_tables}
\title{Check whether you have access to a set of tables}
\usage{
available_tables(update = F)
}
\arguments{
\item{update}{whether to update with new access check or}
}
\description{
Check whether you have access to a set of tables
}
