% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tables.R
\name{check_schema}
\alias{check_schema}
\title{Get the schema of a set of tables}
\usage{
check_schema(tbl_cons)
}
\arguments{
\item{tbl_cons}{table connections from dbplyr}
}
\description{
Get the schema of a set of tables
}
