% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{check_table_access}
\alias{check_table_access}
\title{Check whether you have access to a set of tables}
\usage{
check_table_access(all_tables)
}
\arguments{
\item{all_tables}{set of tables to check for access}
}
\description{
Check whether you have access to a set of tables
}
