% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tables.R
\name{clean_schema}
\alias{clean_schema}
\title{Remove all table items created by loading a schema}
\usage{
clean_schema(con, schema, env = parent.frame())
}
\arguments{
\item{con}{connection to WRDS}

\item{schema}{schema to clean up}
}
\description{
Remove all table items created by loading a schema
}
\details{
This function will remove any tables in the current (parent) environment
that are in the chosen schema. Useful for cleaning up workspaces following
`load_schema()`.
}
