% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_schema}
\alias{explore_schema}
\title{List all of the tables and columns available in a schema}
\usage{
explore_schema(con, schema)
}
\arguments{
\item{con}{connection to wrds}

\item{schema}{schema to explore}
}
\description{
List all of the tables and columns available in a schema
}
