% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{list_columns}
\alias{list_columns}
\title{List columns of a table}
\usage{
list_columns(con, tbl, ...)
}
\arguments{
\item{con}{connection to wrds}

\item{tbl}{name of table}

\item{...}{other parameters to pass to `odbc::odbcListColumns()`}
}
\description{
List columns of a table
}
