% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{list_schemas}
\alias{list_schemas}
\title{List all schemas that match an (optional) pattern}
\usage{
list_schemas(con, pattern = "")
}
\arguments{
\item{con}{connection to database}

\item{pattern}{optional pattern to search across all schemas}
}
\description{
List all schemas that match an (optional) pattern
}
