% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_subscriptions}
\alias{list_subscriptions}
\title{List subscribed data products}
\usage{
list_subscriptions(wrds)
}
\arguments{
\item{wrds}{A \code{DBIConnection} object returned by \code{\link[=wrds_connect]{wrds_connect()}}.}
}
\value{
A tibble with columns \code{schema} and \code{product}.
}
\description{
Returns a tibble of WRDS schemas the user has access to, with
human-readable product names where available.
}
\examples{
\dontrun{
wrds <- wrds_connect()
list_subscriptions(wrds)
wrds_disconnect(wrds)
}
}
