% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tables.R
\name{load_schema}
\alias{load_schema}
\title{Load all tables in a schema into an environment}
\usage{
load_schema(con, schema, env = parent.frame())
}
\arguments{
\item{con}{connection to wrds}

\item{schema}{name of schema to use}

\item{env}{environment to assign these variables in}
}
\description{
Load all tables in a schema into an environment
}
\details{
This is not meant to be used in a production context
where more targeted approaches probably make more sense.
But for users who are relatively new to using SQL and,
for example, want to just load all available crsp tables
into memory, this could be useful. Anyway this is probably bad
practice, and usually annoys me, but in this case I think
it might be useful.
}
