% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{overwrite_in_renviron}
\alias{overwrite_in_renviron}
\title{Overwrite an existing variable in the .Renviron file}
\usage{
overwrite_in_renviron(env_var, newstring)
}
\arguments{
\item{env_var}{environment variable to overwrite}

\item{newstring}{new string to pass as value for environment variable}
}
\description{
Overwrite an existing variable in the .Renviron file
}
