% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{peek}
\alias{peek}
\title{Peek at top n rows of a table}
\usage{
peek(con, tbl, n = 10)
}
\arguments{
\item{con}{connection to wrds}

\item{tbl}{table to query}

\item{n}{number of rows to return}
}
\description{
Peek at top n rows of a table
}
