% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{query}
\alias{query}
\title{Query connection with either a sql script or a string}
\usage{
query(con, q)
}
\arguments{
\item{con}{connection to database}

\item{q}{query to pass along}
}
\description{
Query connection with either a sql script or a string
}
