% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{search_schema}
\alias{search_schema}
\title{Search across schema for matching string}
\usage{
search_schema(con, schema, search, what = "column")
}
\arguments{
\item{con}{connection to wrds}

\item{schema}{schema to search for}

\item{search}{search parameter passed as regular expression}

\item{what}{whether to search for a matching table or column}
}
\description{
Search across schema for matching string
}
\details{
Searches within a schema for specific columns
}
