% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{search_schema_for_column}
\alias{search_schema_for_column}
\title{Search across schema tables for matching column}
\usage{
search_schema_for_column(con, schema, search)
}
\arguments{
\item{con}{connection to wrds}

\item{schema}{schema to search for}

\item{search}{search parameter passed as regular expression}
}
\description{
Search across schema tables for matching column
}
