% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{search_schema_for_table}
\alias{search_schema_for_table}
\title{Search across schema for matching table}
\usage{
search_schema_for_table(con, schema, search)
}
\arguments{
\item{con}{connection to wrds}

\item{schema}{schema to search for}

\item{search}{search parameter passed as regular expression}
}
\description{
Search across schema for matching table
}
