% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{search_wrds}
\alias{search_wrds}
\title{Search WRDS website for documentation}
\usage{
search_wrds(search)
}
\arguments{
\item{search}{string to search for in online WRDS documentation}
}
\description{
Search WRDS website for documentation
}
\details{
This function launches a browser window and searches the WRDS
documentation for the query of interest.
}
