% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{set_wrds_pass}
\alias{set_wrds_pass}
\title{Set password for WRDS connection}
\usage{
set_wrds_pass(pass, ask = T)
}
\arguments{
\item{pass}{username to set and use for connections}

\item{ask}{whether to ask before overwriting existing user id}
}
\description{
Set password for WRDS connection
}
