% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{set_wrds_user}
\alias{set_wrds_user}
\title{Set username for WRDS connection}
\usage{
set_wrds_user(user, ask = T)
}
\arguments{
\item{user}{username to set and use for connections}

\item{ask}{whether to ask before overwriting existing user id}
}
\description{
Set username for WRDS connection
}
