% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrdsClient}
\alias{wrdsClient}
\title{Create a client connection to WRDS}
\usage{
wrdsClient(username = NULL, password = NULL)
}
\arguments{
\item{username}{Username. Don't pass to use stored username.}

\item{password}{Password. Don't pass to use stored password.}
}
\value{
A client connection object
}
\description{
Create a client connection to WRDS
}
\examples{
cl <- wrdsClient("wrdsUser", "xyz123")
cl <- wrdsClient()
}

