% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrdsCredentials}
\alias{wrdsCredentials}
\title{Store your WRDS account credentials
These will be automatically used when calling wrdsClient without parameters
Note: this function does NOT check your credentials!}
\usage{
wrdsCredentials(username, password)
}
\arguments{
\item{username}{Your username}

\item{password}{Your password}
}
\description{
Store your WRDS account credentials
These will be automatically used when calling wrdsClient without parameters
Note: this function does NOT check your credentials!
}
\examples{
wrdsCredentials("wrdsUser", "xyz123")
}

