% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrdsSASPath}
\alias{wrdsSASPath}
\title{Set the SAS drivers path
This is the path that contains sas.core.jar and sas.intrnet.javatools.jar.
The WRDS package needs them to be able to connect to the WRDS service}
\usage{
wrdsSASPath(path)
}
\arguments{
\item{path}{The path to the directory containing the necessary JAR files}
}
\description{
Set the SAS drivers path
This is the path that contains sas.core.jar and sas.intrnet.javatools.jar.
The WRDS package needs them to be able to connect to the WRDS service
}
\examples{
wrdsSASPath("/usr/local/SAS")
}

