% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrdsUpdateTables}
\alias{wrdsUpdateTables}
\title{Updates the link table in the SQLite database
It runs a SQL command on the WRDS cloud, fetches the data, and puts that data in the
SQLite database, in the "link" table}
\usage{
wrdsUpdateTables(conn)
}
\arguments{
\item{conn}{The WRDS cloud connection obtained with wrdsClient()}
}
\description{
Updates the link table in the SQLite database
It runs a SQL command on the WRDS cloud, fetches the data, and puts that data in the
SQLite database, in the "link" table
}
\examples{
cl <- wrdsClient()
wrdsUpdateTables(cl)
}

