% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{list_tables}
\alias{list_tables}
\title{List all tables, optionally a subset within a specific schema}
\usage{
list_tables(con, schema = NULL)
}
\arguments{
\item{con}{connection to database}

\item{schema}{schema to match}
}
\description{
List all tables, optionally a subset within a specific schema
}
