% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrds.R
\name{wrdsUpdateTables}
\alias{wrdsUpdateTables}
\title{Updates the base tables in the SQLite database
It runs SQL commands on the WRDS cloud, fetches the data, and puts that data in their
respective SQLite tables.}
\usage{
wrdsUpdateTables(conn)
}
\arguments{
\item{conn}{The WRDS cloud connection obtained with wrdsClient()}
}
\description{
Updates the base tables in the SQLite database
It runs SQL commands on the WRDS cloud, fetches the data, and puts that data in their
respective SQLite tables.
}
\examples{
cl <- wrdsClient()
wrdsUpdateTables(cl)
}

